+++ web/ui/react-app/src/pages/serviceDiscovery/LabelsTable.tsx
@@ -1,15 +1,13 @@
-import React, { FC, useEffect, useState } from 'react';
+import React, { FC, useState } from 'react';
 import { Badge, Table } from 'reactstrap';
 import { TargetLabels } from './Services';
 import { ToggleMoreLess } from '../../components/ToggleMoreLess';
-import InfiniteScroll from 'react-infinite-scroll-component';
+import CustomInfiniteScroll, { InfiniteScrollItemsProps } from '../../components/CustomInfiniteScroll';
 
 interface LabelProps {
   value: TargetLabels[];
   name: string;
 }
-
-const initialNumberOfTargetsDisplayed = 50;
 
 const formatLabels = (labels: Record<string, string> | string) => {
   return Object.entries(labels).map(([key, value]) => {
@@ -23,26 +21,35 @@ const formatLabels = (labels: Record<string, string> | string) => {
   });
 };
 
+const LabelsTableContent: FC<InfiniteScrollItemsProps<TargetLabels>> = ({ items }) => {
+  return (
+    <Table size="sm" bordered hover striped>
+      <thead>
+        <tr>
+          <th>Discovered Labels</th>
+          <th>Target Labels</th>
+        </tr>
+      </thead>
+      <tbody>
+        {items.map((_, i) => {
+          return (
+            <tr key={i}>
+              <td>{formatLabels(items[i].discoveredLabels)}</td>
+              {items[i].isDropped ? (
+                <td style={{ fontWeight: 'bold' }}>Dropped</td>
+              ) : (
+                <td>{formatLabels(items[i].labels)}</td>
+              )}
+            </tr>
+          );
+        })}
+      </tbody>
+    </Table>
+  );
+};
+
 export const LabelsTable: FC<LabelProps> = ({ value, name }) => {
   const [showMore, setShowMore] = useState(false);
-  const [items, setItems] = useState<TargetLabels[]>(value.slice(0, 50));
-  const [index, setIndex] = useState<number>(initialNumberOfTargetsDisplayed);
-  const [hasMore, setHasMore] = useState<boolean>(value.length > initialNumberOfTargetsDisplayed);
-
-  useEffect(() => {
-    setItems(value.slice(0, initialNumberOfTargetsDisplayed));
-    setHasMore(value.length > initialNumberOfTargetsDisplayed);
-  }, [value]);
-
-  const fetchMoreData = () => {
-    if (items.length === value.length) {
-      setHasMore(false);
-    } else {
-      const newIndex = index + initialNumberOfTargetsDisplayed;
-      setIndex(newIndex);
-      setItems(value.slice(0, newIndex));
-    }
-  };
 
   return (
     <>
@@ -56,38 +63,7 @@ export const LabelsTable: FC<LabelProps> = ({ value, name }) => {
           <span className="target-head">{name}</span>
         </ToggleMoreLess>
       </div>
-      {showMore ? (
-        <InfiniteScroll
-          next={fetchMoreData}
-          hasMore={hasMore}
-          loader={<h4>loading...</h4>}
-          dataLength={items.length}
-          height={items.length > 25 ? '75vh' : ''}
-        >
-          <Table size="sm" bordered hover striped>
-            <thead>
-              <tr>
-                <th>Discovered Labels</th>
-                <th>Target Labels</th>
-              </tr>
-            </thead>
-            <tbody>
-              {items.map((_, i) => {
-                return (
-                  <tr key={i}>
-                    <td>{formatLabels(items[i].discoveredLabels)}</td>
-                    {items[i].isDropped ? (
-                      <td style={{ fontWeight: 'bold' }}>Dropped</td>
-                    ) : (
-                      <td>{formatLabels(items[i].labels)}</td>
-                    )}
-                  </tr>
-                );
-              })}
-            </tbody>
-          </Table>
-        </InfiniteScroll>
-      ) : null}
+      {showMore ? <CustomInfiniteScroll value={value} child={LabelsTableContent} /> : null}
     </>
   );
 };
